+++ api-config.js
@@ -4,7 +4,7 @@ module.exports = {
     path: 'src/packages/application/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/application',
       },
     ],
@@ -12,22 +12,24 @@ module.exports = {
   // base
   env: {
     path: 'src/packages/base/env/src/index.ts',
-    // 是否可以按环境拆包
     unNeedSplit: true,
+    // 是否需要公共utils
+    needCommonUtil: false,
     pkgInfo: [
       {
-        version: '1.0.1',
+        version: '1.0.2',
         name: '@uni/env',
       },
     ],
   },
   canIUse: {
     path: 'src/packages/base/canIUse/src/index.ts',
-    // 是否可以按环境拆包
     unNeedSplit: true,
+    // 是否需要公共utils
+    needCommonUtil: false,
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/caniuse',
       },
     ],
@@ -36,12 +38,14 @@ module.exports = {
   unitTool: {
     path: 'src/packages/base/unit-tool/src/index.ts',
     unNeedSplit: true,
+    // 是否需要公共utils
+    needCommonUtil: false,
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/unit-tool',
         dependencies: {
-          '@uni/system-info': '^1.0.2',
+          '@uni/system-info': '^1.0.3',
         },
       },
     ],
@@ -51,7 +55,7 @@ module.exports = {
     path: 'src/packages/canvas/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/canvas',
       },
     ],
@@ -61,7 +65,7 @@ module.exports = {
     path: 'src/packages/device/accelerometer/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/accelerometer',
       },
     ],
@@ -70,7 +74,7 @@ module.exports = {
     path: 'src/packages/device/clipboard/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/clipboard',
       },
     ],
@@ -79,7 +83,7 @@ module.exports = {
     path: 'src/packages/device/systemInfo/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/system-info',
       },
     ],
@@ -98,7 +102,7 @@ module.exports = {
     path: 'src/packages/file/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/file',
       },
     ],
@@ -108,7 +112,7 @@ module.exports = {
     path: 'src/packages/interactive/actionSheet/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/action-sheet',
       },
     ],
@@ -117,10 +121,10 @@ module.exports = {
     path: 'src/packages/interactive/alert/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/alert',
         dependencies: {
-          '@uni/confirm': '^1.0.2',
+          '@uni/confirm': '^1.0.3',
         },
       },
     ],
@@ -157,7 +161,7 @@ module.exports = {
     path: 'src/packages/interactive/toast/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.3',
+        version: '1.0.4',
         name: '@uni/toast',
       },
     ],
@@ -166,7 +170,7 @@ module.exports = {
     path: 'src/packages/interactive/loading/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/loading',
       },
     ],
@@ -184,7 +188,7 @@ module.exports = {
     path: 'src/packages/interactive/element/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/element',
       },
     ],
@@ -193,7 +197,7 @@ module.exports = {
     path: 'src/packages/interactive/intersectionObserver/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/intersection-observer',
       },
     ],
@@ -202,7 +206,7 @@ module.exports = {
     path: 'src/packages/interactive/confirm/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/confirm',
       },
     ],
@@ -220,7 +224,7 @@ module.exports = {
     path: 'src/packages/location/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/location',
       },
     ],
@@ -230,7 +234,7 @@ module.exports = {
     path: 'src/packages/media/image/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/image',
       },
     ],
@@ -240,7 +244,7 @@ module.exports = {
     path: 'src/packages/navigate/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/navigate',
       },
     ],
@@ -250,7 +254,7 @@ module.exports = {
     path: 'src/packages/network/request/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/request',
       },
     ],
@@ -260,7 +264,7 @@ module.exports = {
     path: 'src/packages/storage/src/index.ts',
     pkgInfo: [
       {
-        version: '1.0.2',
+        version: '1.0.3',
         name: '@uni/storage',
       },
     ],