+++ scripts/build-plugin/build.js
@@ -21,7 +21,8 @@ module.exports = (api ,options = {}) => {
     buildPkgJson: true,
     mvReadme: true,
     mvNpmrc: true,
-    declaration: true
+    declaration: true,
+    needCommonUtil: true
   };
   let apiInfo;
   let sourceMap;
@@ -39,14 +40,9 @@ module.exports = (api ,options = {}) => {
         name: mainPkg.name,
       };
       output = outputDir + 'main/';
-      apiInfo = {...defaultApiInfo, unNeedSplit: false};
+      apiInfo = {...defaultApiInfo, needCommonUtil: false};
 
       buildMain(rootDir, sourceMap);
-      taskList.push(async function(ctx, next) {
-        await initPkg(entry, pkgInfo, output, sourceMap, apiInfo, isMain);
-        gulpRelease(api, { entry, output, isMain });
-        await next();
-      });
     } else {
       entry = path.resolve(rootDir, sourceMap[apiName].path);
       apiInfo = {...defaultApiInfo, ...sourceMap[apiName]};
@@ -55,20 +51,10 @@ module.exports = (api ,options = {}) => {
         sourceMap[apiName].pkgInfo.forEach(i => {
           pkgInfo = i;
           output = outputDir + getItemOutputPath(i.name, sourceMap[apiName].outputDir);
-          taskList.push(async function(ctx, next) {
-            await initPkg(entry, pkgInfo, output, sourceMap, apiInfo, isMain);
-            gulpRelease(api, { entry, output, isMain });
-            await next();
-          });
         });
       } else {
         pkgInfo = sourceMap[apiName].pkgInfo;
         output = outputDir + getItemOutputPath(sourceMap[apiName].pkgInfo.name, sourceMap[apiName].outputDir);
-        taskList.push(async function(ctx, next) {
-          await initPkg(entry, pkgInfo, output, sourceMap, apiInfo, isMain);
-          gulpRelease(api, { entry, output, isMain });
-          await next();
-        });
       }
     }
     
@@ -81,14 +67,9 @@ module.exports = (api ,options = {}) => {
       mvNpmrc: options.npmrc,
       mvReadme: options.readme,
       declaration: options.declaration,
-      unNeedSplit: options.unNeedSplit,
+      needCommonUtil: false,
     };
     sourceMap = null;
-    taskList.push(async function(ctx, next) {
-        await initPkg(entry, pkgInfo, output, sourceMap, apiInfo, isMain);
-        gulpRelease(api, { entry, output, isMain });
-        await next();
-    });
   }
   rm.sync(path.resolve(rootDir, output));
   webpackRelease(api, {
@@ -100,21 +81,11 @@ module.exports = (api ,options = {}) => {
     isMain
   });
   
-  // onHook('before.build.load', async () => {
-  //   // consoleClear(true);
-  //   await initPkg(entry, pkgInfo, output, sourceMap, apiInfo, isMain);
-  // });
+  onHook('before.build.load', async () => {
+    // consoleClear(true);
+    await initPkg(entry, pkgInfo, output, sourceMap, apiInfo, isMain);
+  });
   onHook('after.build.compile', () => {
-    // if(!isMain) {
-      // gulpRelease(api, { entry, output, isMain });
-    // } else {
-    //   rollupRelease(entry, pkgInfo, output, sourceMap, apiInfo, isMain)
-    // }
-    
-    compose(taskList)().then(function() {
-      // logger('END', {status: 'SUCCESS'});
-    }).catch(function(err) {
-      console.log(err);
-    });
+    gulpRelease(api, { entry, output, isMain, apiInfo });
   });
 }