+++ src/packages/interactive/toast/src/wechat-miniprogram/hideToast.ts
@@ -1,0 +1,8 @@
+import { normalize } from '../utils/index';
+import { HideToastOption } from '../types';
+
+export const hide = normalize((options?: HideToastOption): void => {
+  wx.hideToast(options);
+});
+
+export default hide;